Rename AC module
